use common ros and gazebo functions
