--- eufs_plugins/gazebo_race_car_plugin/include/gazebo_race_car_plugin/helpers_ros.hpp
@@ -1,7 +1,0 @@
-#include "rclcpp/rclcpp.hpp"
-
-bool is_initalised(rclcpp::PublisherBase::SharedPtr publisher) { return (bool)publisher; }
-
-bool has_subscribers(rclcpp::PublisherBase::SharedPtr publisher) {
-    return is_initalised(publisher) && publisher->get_subscription_count() > 0;
-}
